[doc] Block diagram added
This diagram is generated with D2 so there is a readme to describe how
to generate this. It also includes an attribute file to say that the SVG
is auto-generated.

The diagram shows which blocks are cached an uncached as well as which
blocks are tag aware and tag oblivious.